% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP.R
\name{CP}
\alias{CP}
\title{CP}
\usage{
CP(bw, dg)
}
\arguments{
\item{bw}{body weight (kg).}

\item{dg}{daily gain (kg).}
}
\description{
This function calculate the crude protein for maintain and growth of Hanwoo. CP = NP/eP
}
\examples{
CP(bw=150, dg=1)
}
\keyword{Hanwoo}
