% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPm.R
\name{CPm}
\alias{CPm}
\title{CPm}
\usage{
CPm(bw)
}
\arguments{
\item{bw}{body weight (kg).}
}
\description{
This function calculate the crude protein for maintain of Hanwoo. CPm = 5.56 * bw^0.75
}
\examples{
CPm(bw=150)
}
\keyword{Hanwoo}
