% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanwoo1}
\alias{hanwoo1}
\title{Hanwoo body weight dataset (set 1).}
\format{
A data frame with body weight measurements for Hanwoo cattle
}
\usage{
hanwoo1
}
\description{
A dataset of body weight records for Hanwoo cattle.
}
\keyword{datasets}
