% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanwoo2}
\alias{hanwoo2}
\title{Hanwoo body weight dataset (set 2).}
\format{
A data frame with body weight measurements for Hanwoo cattle
}
\usage{
hanwoo2
}
\description{
A dataset of body weight records for Hanwoo cattle.
}
\keyword{datasets}
