% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_bull.R
\name{hanwoo_bull}
\alias{hanwoo_bull}
\title{hanwoo_bull}
\usage{
hanwoo_bull(KPN, type = "list")
}
\arguments{
\item{KPN}{KPN number of a bull you get the inform.}

\item{type}{"list", "all", or "selected"}
}
\value{
Depending on the type parameter:
  \describe{
    \item{list}{A list with two elements: 'All' (full information) and 'EPD' (selected EPD values)}
    \item{all}{A tibble with all available bull information}
    \item{selected}{A tibble with selected EPD information (kpn, father, selling, guarantee, birthday, carcass_weight_kg, longissimus_cm, backfat_mm, marbling)}
  }
}
\description{
This function scraping the bull information of Hanwoo from data.go.kr.
}
\examples{
\dontrun{
hanwoo_bull(KPN = 950)
hanwoo_bull(KPN = 950, type = "selected")
}
}
\keyword{Hanwoo}
\keyword{bull}
