% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanwoo_farm}
\alias{hanwoo_farm}
\title{Distribution of average grade of farms in South Korea}
\format{A list with 3 types of sex (steer, heifer, and cow)}
\usage{
hanwoo_farm
}
\description{
Last updated: 2021-12-08
}
\keyword{datasets}
