% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_key.R
\name{hanwoo_key}
\alias{hanwoo_key}
\title{hanwoo_key}
\usage{
hanwoo_key(key)
}
\arguments{
\item{key}{An API key that you get from data.go.kr.}
}
\description{
A function for set your API key.
}
\examples{
hanwoo_key(key="1KTzXwIIp1Gn1weD3eRnsmoLgP4efuf\%2FH27OczlrUCFKKHRYxk4Rn9OXxJpTlNd\%2BU\%2Fw\%3D\%3D_PLESE_GET_YOUR_API_KEY_AT_DATA.GO.KR")
}
\keyword{Hanwoo}
