% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hanwoo_kpn}
\alias{hanwoo_kpn}
\title{Korean Proven Bull (KPN) dataset.}
\format{A data frame with 1,599 rows and 134 variables}
\usage{
hanwoo_kpn
}
\description{
This dataset contains a dataset of the Korean Proven Bulls
from National Institute of Animal Science.
}
\keyword{datasets}
