% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_stock.R
\name{hanwoo_stock}
\alias{hanwoo_stock}
\title{hanwoo_stock}
\usage{
hanwoo_stock()
}
\value{
A tibble containing Hanwoo stock data with columns: date, price information,
  year, week, and wday (day of week).
}
\description{
This function get the stock calf and beef price of hanwoo. Data from ekapepia.com.
}
\examples{
\dontrun{
hanwoo_stock()
}
}
\keyword{Hanwoo}
