% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kcalendar}
\alias{Kcalendar}
\title{Lunar Calendar of Korea.}
\format{
A data frame with 241,027 rows and 11 variables
}
\usage{
Kcalendar
}
\description{
This dataset contains a subset of the Lunar calendar of Korea.
}
\keyword{datasets}
