% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ME.R
\name{ME}
\alias{ME}
\title{ME}
\usage{
ME(bw, dg)
}
\arguments{
\item{bw}{body weight.}

\item{dg}{daily gain (kg).}
}
\description{
This function calculate the metabolizable energy for maintain of Hanwoo.
}
\examples{
ME(bw=150, dg=1)
}
\keyword{Hanwoo}
