% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEg.R
\name{MEg}
\alias{MEg}
\title{MEg}
\usage{
MEg(bw, dg)
}
\arguments{
\item{bw}{body weight (kg).}

\item{dg}{daily gain (kg).}
}
\description{
This function calculate the net energy for growth of Hanwoo.
}
\examples{
MEg(bw=150,dg=0.8)
}
\keyword{Hanwoo}
