% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEm.R
\name{MEm}
\alias{MEm}
\title{MEm}
\usage{
MEm(bw)
}
\arguments{
\item{bw}{body weight.}
}
\description{
This function calculate the metabolizable energy for maintain of Hanwoo.
}
\examples{
MEm(bw=150)
}
\keyword{Hanwoo}
