% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEg.R
\name{NEg}
\alias{NEg}
\title{NEg}
\usage{
NEg(bw, dg)
}
\arguments{
\item{bw}{body weight (kg).}

\item{dg}{daily gain (kg).}
}
\description{
This function calculate the net energy for growth of Hanwoo. 
NEg = 0.0533 \* body weight^0.75 \* daily gain (unit=kg)
}
\examples{
NEg(bw=150,dg=0.8)
}
\keyword{Hanwoo}
