% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP.R
\name{NP}
\alias{NP}
\title{NP}
\usage{
NP(bw, dg)
}
\arguments{
\item{bw}{body weight (kg).}

\item{dg}{daily gain (kg).}
}
\description{
This function calculate the net protein for maintain and growth of Hanwoo. NP = NPm + NPg
}
\examples{
NP(bw=150, dg=1)
}
\keyword{Hanwoo}
