% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPg.R
\name{NPg}
\alias{NPg}
\title{NPg}
\usage{
NPg(dg, bw)
}
\arguments{
\item{dg}{daily gain (kg).}

\item{bw}{body weight (kg).}
}
\description{
This function calculate the net protein for growth (NPg = RP) of Hanwoo. NPg = dg \* (224.7 - 0.251 \* bw)
}
\examples{
NPg(bw=150, dg=1)
}
\keyword{Hanwoo}
