% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPm.R
\name{NPm}
\alias{NPm}
\title{NPm}
\usage{
NPm(bw)
}
\arguments{
\item{bw}{body weight (kg).}
}
\description{
This function calculate the net protein for maintain of Hanwoo. NPm = CPm * 0.51
}
\examples{
NPm(bw=150)
}
\keyword{Hanwoo}
