% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{program_nias}
\alias{program_nias}
\title{Nutritional guide program for the Hanwoo steers.}
\format{A data frame with 24 rows and 10 variables
\describe{
  \item{stage3}{3 step program for steer}
  \item{stage4}{4 step program for steer}
  \item{month}{Month of hanwoo steer}
  \item{weight}{Unshrunk body weight of hanwoo steer, kg}
  \item{adg}{Average daily gain}
  \item{feed.intake}{Feed intake, dry matter basis, kg}
  \item{max.git}{Maximum growth stage for gastro intestinal tract}
  \item{max.muscle}{Maximum growth stage for muscle}
  \item{max.weight}{Maximum growth stage for body weight}
  \item{max.fat}{Maximum growth stage for intra-muscular fat}
}}
\usage{
program_nias
}
\description{
This dataset contains a subset of the Nutritional guide program
(National Institute of Animal Science)for the Hanwoo.
Body weight,month, average daily gain, and maximum growth
information dataset are available.
}
\keyword{datasets}
