% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_bull.R
\name{req_bull}
\alias{req_bull}
\title{req_bull}
\format{
\describe{
  \item{BW}{Body weight, kg}
  \item{ADG}{Avergae daily gain, kg}
  \item{CP}{Crude protein}
  \item{ME}{Metabolizable energy}
  \item{NEg}{Net energy for growth}
  \item{Ca}{Calcium}
  \item{P}{Phosphorus}
  \item{vit}{Vitamin}
}
}
\usage{
req_bull(bw, dg)
}
\arguments{
\item{bw}{Body weight, kg}

\item{dg}{Daily gain, kg}
}
\value{
A tibble with 10 columns containing nutrient requirements:
  sex, BW_kg, ADG_kg, CP_g, ME_Mcal, NEg_Mcal, Ca_g, P_g,
  vitA_1000IU, and vitD_1000IU.
}
\description{
Calculate the nutrient requirement of Hanwoo bull.
}
\examples{
req_bull(bw = 150, dg = 0.8)
}
\keyword{Hanwoo}
\keyword{Nutrient}
\keyword{requirement}
