% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steer_c}
\alias{steer_c}
\title{Body weight prediction model for Hanwoo steers using chest girth.}
\format{
An object of class \code{lm}
\describe{
  \item{chest}{Chest girth of the steer, cm}
}
}
\usage{
steer_c
}
\description{
A linear regression model for predicting body weight (kg) from
chest girth (cm) of Hanwoo steers.
}
\keyword{datasets}
