% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steer_h}
\alias{steer_h}
\title{Body weight prediction model for Hanwoo steers using body height.}
\format{
An object of class \code{lm}
\describe{
  \item{height}{Body height of the steer, cm}
}
}
\usage{
steer_h
}
\description{
A linear regression model for predicting body weight (kg) from
body height (cm) of Hanwoo steers.
}
\keyword{datasets}
