% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steer_m}
\alias{steer_m}
\title{Body weight prediction model for Hanwoo steers using age.}
\format{
An object of class \code{lm}
\describe{
  \item{month}{Age of the steer, months}
}
}
\usage{
steer_m
}
\description{
A linear regression model for predicting body weight (kg) from
age in months of Hanwoo steers.
}
\keyword{datasets}
