% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steer_mhlc}
\alias{steer_mhlc}
\title{Body weight prediction model for Hanwoo steers using multiple measurements.}
\format{
An object of class \code{lm}
\describe{
  \item{month}{Age of the steer, months}
  \item{height}{Body height of the steer, cm}
  \item{length}{Body length of the steer, cm}
  \item{chest}{Chest girth of the steer, cm}
}
}
\usage{
steer_mhlc
}
\description{
A linear regression model for predicting body weight (kg) from
month (age), body height, body length, and chest girth of Hanwoo steers.
}
\keyword{datasets}
