% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitA.R
\name{vitA}
\alias{vitA}
\title{vitA}
\usage{
vitA(bw)
}
\arguments{
\item{bw}{body weight (kg).}
}
\description{
This function calculate the vitamin A requirement of Hanwoo. Vitamin A = 0.0424 * body weight
}
\examples{
vitA(bw=150)
}
\keyword{Hanwoo}
