% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_ku}
\alias{weight_ku}
\title{Body weight dataset for the Hanwoo steers in KU farm.}
\format{A data frame with 780 rows and 3 variables
\describe{
  \item{treat}{diet treatment}
  \item{month}{month of hanwoo steer}
  \item{weight}{Unshrunk body weight of hanwoo steer, kg}
}}
\usage{
weight_ku
}
\description{
This dataset contains a subset of the body weight dataset for
Hanwoo steer in Konkuk University farm.
}
\keyword{datasets}
