% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_bull.R
\name{hanwoo_bull}
\alias{hanwoo_bull}
\title{hanwoo_bull}
\usage{
hanwoo_bull(KPN, type = "list")
}
\arguments{
\item{KPN}{KPN number of a bull you get the inform.}

\item{type}{"list", "all", or "selected"}
}
\description{
This function scraping the bull information of Hanwoo from data.go.kr.
}
\examples{
hanwoo_bull(KPN = 950)
hanwoo_bull(KPN = 950, type = "selected")
}
\keyword{Hanwoo}
\keyword{bull}
