% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_bull.R
\name{hanwoo_bull}
\alias{hanwoo_bull}
\title{hanwoo_bull}
\usage{
hanwoo_bull(KPN)
}
\arguments{
\item{KPN}{KPN number of a bull you get the inform.}
}
\description{
This function scraping the bull information of Hanwoo from data.go.kr.
}
\examples{
hanwoo_bull(KPN=1080)
}
\keyword{Hanwoo}
\keyword{bull}
