% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_info.R
\name{hanwoo_info}
\alias{hanwoo_info}
\title{hanwoo_info}
\usage{
hanwoo_info(cattle, key_encoding, key_decoding, time_check = FALSE)
}
\arguments{
\item{cattle}{Number of cattle to get the information for.}

\item{key_encoding}{Encoded API key from data.go.kr.}

\item{key_decoding}{Decoded API key from data.go.kr.}

\item{time_check}{Check the response time of the server.}
}
\description{
This function scrapes the information of Hanwoo from data.go.kr. Please get your API key and request for application at data.go.kr.
}
\examples{
hanwoo_info(cattle = "002083191603", key_encoding, key_decoding)
hanwoo_info(cattle = "002115280512", key_encoding, key_decoding)
}
\keyword{Hanwoo}
