% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_info.R
\name{hanwoo_info}
\alias{hanwoo_info}
\title{hanwoo_info}
\usage{
hanwoo_info(cattle)
}
\arguments{
\item{cattle}{Number of cattle you get the inform.}
}
\description{
This function scraping the information of Hanwoo from data.go.kr. Please get your API key and request for applicate at data.go.kr.
}
\examples{
hanwoo_info(cattle="002083191603")
}
\keyword{Hanwoo}
