% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_price.R
\name{hanwoo_price}
\alias{hanwoo_price}
\title{hanwoo_price}
\usage{
hanwoo_price(date = "", type = "df")
}
\arguments{
\item{date}{date you get the inform. e.g. 20190510}
}
\description{
This function scrap the real-time (within 5 days) price information of Hanwoo from data.go.kr. Please get your API key and request for applicate at data.go.kr.
}
\examples{
hanwoo_price(date = "", type = "list")
hanwoo_price(date = "20190510", type = "df")
}
\keyword{Hanwoo}
