% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo_stock.R
\name{hanwoo_stock}
\alias{hanwoo_stock}
\title{hanwoo_stock}
\usage{
hanwoo_stock(by = "summary")
}
\description{
This function get the stock calf and beef price of hanwoo. Data from ekapepia.com. It needs the Google OAuth access credentials.
}
\examples{
hanwoo_stock()
}
\keyword{Hanwoo}
