% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_steer.R
\name{req_steer}
\alias{req_steer}
\title{req_steer}
\usage{
req_steer(bw, dg)
}
\arguments{
\item{bw}{Body weight}

\item{dg}{Daily gain}
}
\description{
Calculate the nutrient requirement of Hanwoo steer.
}
\examples{
hanwoo_info(bw = 150, dg = 0.8)
}
\keyword{Hanwoo}
\keyword{Nutrient}
\keyword{requirement}
