% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_steer.R
\name{req_steer}
\alias{req_steer}
\title{req_steer}
\format{
\describe{
  \item{BW}{Body weight, kg}
  \item{ADG}{Avergae daily gain, kg}
  \item{CP}{Crude protein}
  \item{ME}{Metabolizable energy}
  \item{NEg}{Net energy for growth}
  \item{Ca}{Calcium}
  \item{P}{Phosphorus}
  \item{vit}{Vitamin}
}
}
\usage{
req_steer(bw, dg)
}
\arguments{
\item{bw}{Body weight}

\item{dg}{Daily gain}
}
\description{
Calculate the nutrient requirement of Hanwoo steer.
}
\examples{
req_steer(bw = 150, dg = 0.8)
}
\keyword{Hanwoo}
\keyword{Nutrient}
\keyword{requirement}
