% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitD.R
\name{vitD}
\alias{vitD}
\title{vitD}
\usage{
vitD(bw)
}
\arguments{
\item{bw}{body weight (kg).}
}
\description{
This function calculate the vitamin A requirement of Hanwoo. Vitamin D = 0.0424 * body weight
}
\examples{
vitD(bw=150)
}
\keyword{Hanwoo}
